+++ Altinn/AT.Common.Altinn.Publish/Model/Adapter/AltinnInstanceSummary.cs
@@ -4,12 +4,12 @@ using Arbeidstilsynet.Common.Altinn.Model.Api.Response;
 namespace Arbeidstilsynet.Common.Altinn.Model.Adapter;
 
 /// <summary>
-/// A summary of an Altinn AltinnInstance. An Altinn AltinnInstance represents a submitted form in Altinn.
+/// A summary of an Altinn instance, which represents a submitted form in Altinn.
 /// </summary>
 public record AltinnInstanceSummary
 {
     /// <summary>
-    /// Metadata about the AltinnInstance.
+    /// Metadata about the Altinn instance.
     /// </summary>
     public required AltinnMetadata Metadata { get; init; }
 
@@ -19,7 +19,7 @@ public record AltinnInstanceSummary
     public required AltinnDocument SkjemaAsPdf { get; init; }
 
     /// <summary>
-    /// The structured data of the AltinnInstance. This is identified by <see cref="AltinnAppSpecification.StructuredDataTypeId"/>
+    /// The structured data of the Altinn instance. This is identified by <see cref="AltinnAppSpecification.StructuredDataTypeId"/>
     /// </summary>
     public AltinnDocument? StructuredData { get; init; }